requirements.txt
django
djangorestframework
environs
djangorestframework-simplejwt
psycopg2
pandas
celery
rabbitmq
